Create test plan when new swarm runner is started
Modify reporter script, so it can parse results of
system tests that are running at the moment.
Also if test failed try to find link to Launchpad
bug in previous results for the same test case. Copy
link to bug only if it's targeted to given project
and milestone and its status isn't 'Fix Released'.
In case linked bug is marked as duplicate of another
bug, check the status of that bug. If linked bug is
private or doesn't exist skip it.

Change-Id: I2392c81bcad6273ecbabfee479b74263fc7c0fbf
Closes-bug: #1419865