Split failover tests to separate cases for nova and neutron
We have the same test cases for nova and neutron in test_failover
that are started in different threads with different settings.

Test cases remain the same but are called from different test
methods (separately for nova and for neutron) to provide
dependency on different environments and snapshots.

test_failover.py separated into test_failover.py and
test_failover_base.py for more suitable developing.

Change-Id: Ib0a01b621c1a0b5d8ed3eadbc21262509563ddc4
Closes-Bug: #1422295