Fixing issue with logging in method "assert_ostf_run_certain"
The method "assert_ostf_run_certain" takes a list of tests, that
must have success, as one of the parameters. If the list has more
than 1 item, logging in the method "assert_ostf_run_certain" will
work incorrectly. For example, if we have three tests in the list,
only the last item of this list will always be passed into function
for logging.

Closes-Bug: #1433184

Change-Id: I9887dee114c972c4b69823d794a491728f112ebc