"update_node_networks" method was updated for Contrail plugin and "Install plugin" test case was added
In the original implementation it is impossible to execute tests for Contrail plugin. For "update_node_networks"
method was added boolean flag that specifies whether to use standard method for updating data or specific for Contrail tests.
Added methods for Contrail plugin installation with the following scenario:
1. Upload plugin to the master node
2. Install plugin with VLAN segmentation
Added import in run_tests.py for running Contrail test cases and updated settings.py file

Change-Id: Id05a77d368d32a4ca8645a759111d7a00d0fba9b