Log waiting time for node bootstrapping
We had a staging tests failure (see the bug referenced below) which was
caused by 3 of 5 nodes not comming up in time. Right now we don't know
how much time it takes 5 nodes to come up normally. Maybe 10 minutes is
barely enough and we just need to increase timeout? We will not know that
for sure until we log the time and see how much time it takes on average.
This change introduces such logging.

Change-Id: I342f6997f74d2d42dcf772fb5e2f597122d54a94
Related-Bug: #1435837