Add explicit timeout to deploy_cluster_wait
Before this we use self.fuel_web.deploy_cluster_wait(cluster_id) for
deploying a cluster. See, there are not timeout because default timeout
for assert_task_success was enough for deploying. But now it isn't.
Some tests works too long and breaks by timeout. And I add possibility
to set this timeout explicitly when it needed.

Change-Id: Id3118a00637e88e018077e647f2c13d8883bf0eb