Add waiting for keystone to be fully available after haproxy restart
Sometimes keystone is not available right after haproxy restart. To
consider that in the tests preliminary check of the situation added in
form of ostf sanity set execution inside of try...catch and testing
exception type if any. In case when the check is successful, sleep for
two minutes (approximate time for the service getting available)
is performed after which main bulk of ostf test sets are executed.

Change-Id: I4a5d6adcc13ca1e996e0ed6bdbaa0ef35352f4b9
Partial-Bug: #1446590