Add waiting for keystone to be fully available after haproxy restart
Sometimes keystone is not available right after haproxy restart. To
consider that in the tests preliminary check of the situation added in
form of ostf sanity set execution. The tests are constantly rerunning in
given timeout period in case of failing with expected exception. The
timeout value is 180 sec. (approximate time for the service getting
available). After preliminary checking completion main bulk of ostf tests are
executed.

Change-Id: I4a5d6adcc13ca1e996e0ed6bdbaa0ef35352f4b9
Partial-Bug: #1446590