Get rid of network bonds with 'balance-*' modes
Since it was decided to move 'balance-tlb' and 'balance-alb'
bonding modes to experimental features (see bug #1458935),
there is no need to run tests with them as they expectedly
fail. But we still need to check bonding feature with all
supported network types, so switch bonding mode to
'active-backup' in all tests.

Change-Id: Id75f8b17a28dc8da27ca4b4ff8e3ceb1fee69d83
Partial-bug: #1319413