Use list of 'affected-pkgs' from erratum in tests
Since it's almost impossible to know fixed package
version until the fix is merged, there is no sense
to list the same packages in two sections 'affected-pkgs'
and 'fixed-pkgs'. The first one contains versions of
affected packages so it should be kept, and the second
one became redundant (because versions aren't specified
there), so it is going to be removed from erratum.yaml.

Change-Id: I09614f2970ea911ba125088cf658e6da33bc8382
Implements: blueprint mos-patching-tests