Avoiding an extra gap in the name of a test plan
In fuel-qa/fuelweb_test/testrail/report.py we have the following code
block http://paste.mirantis.net/show/769/. If the prefix variable is '',
the name of a test plan will have an extra gap. So other scripts (for
example, report_tempest_results.py), that publish some test results to
TestRail, always create new test plan because they try to find the name
of the test plan without the gap. That causes duplicating test plans and
that is very inconvenient for users.

Closes-Bug: #1476327

Change-Id: I1c087e835e63db4828fafb3b936f7ace2d4eb29d