Fixed logic for tempest tests reporting
For now we don't use In Progress status during the tests execution
and also we use Failed status if deployment of master node will fail.
We need to change this logic and start to change status of test cases
to In Progress during the tests execution and Blocked in case of
some issues during master node failures or cloud deployment issues.

Closes-Bug: #1487470

Change-Id: Id1112f470c6470e0cb741a5327511d49db8f548c