Improve 'run_on_remote' for more general use
To use method in cases when ec is expected to be not zero and reduce
duplicated code in asserts messages new parameters in 'run_on_remote'
method proposed with following logic:
* 'assert_ec_equal' with default '[0]'. Error message will be logged only
when exit_code != assert_ec_equal. Can pass xrange(1,8) or [1, 4, 5]
* 'raise_on_assert' with default 'True'. Exception now can be
surpressed when exit_code != assert_ec_equal if 'raiseOnAssert' is 'False'

Also:
* jsonify parameter removed. New method 'json_deserialize' created
instead. Input type is string or list.
* 'run_on_remote_get_results' return dict with:
  'command',
  'host',
  'stdout',
  'stdout_str',
  'stdout_len',
  'stderr',
  'stderr_str',
  'stderr_len',
  'exit_code', instead of just stdout string
* 'run_on_remote' wrapper leaved for compatibility

Change-Id: I32a9d743929d01857a88844976074df719d9880a
Closes-Bug: #1489165