Increase timeout for env deletion over cli
In test delete cluster over cli we use
revert mechanism, but do not check if all the nodes
become online after revert. This lead to the situation
when cluster deletion takes a little bit more time.
So In this patch delete timeout was increased.
Also add check for node statuses after revert
and log with error level if some of them is offline.
We don't interrupt tests here according to cluster should be deleted,even all the nodes are offline

Change-Id: I769e5a17875595b632d111d6b5f1a782989e3a37
Closes-Bug: #1508007