Added execution of MOS tests to SWARM
We have a lot of integration tests for MOS verification
in upstream and in separate repository
https://github.com/Mirantis/mos-integration-tests.
We are going to move all our tests to one jenkins with
SWARM tests and this is why we need to add system tests
which will execute all integration tests in SWARM test suite.

These tests will be executed on the separate environment
as well as Tempest tests or integration tests for Sahara
and Murano deployments validation.

We need to add SWARM tests which will export all required
variables and will run run_tests.sh script from
https://github.com/Mirantis/mos-integration-tests.

Change-Id: I9ed6625a2a57a9aa6f2b65d9b3cf94894c70519b
Closes-Bug: #1523436