Add gate test for Astute
Add gate test for Astute that should be executed for each review in openstack/fuel-astute repo.

Next variables should be exported:
UPDATE_FUEL=True
UPDATE_FUEL_PATH=/path/to/rpm/to/be/downloaded

Test depends on ready_with_3_slaves, so it is better to run it with -k/-K options.

Scenario:

  1. Revert environment ready_with_3_slaves
  2. Update Astute rpm package from review
  3. Update network configuration
  4. Create env
  5. Update nodes with roles: controller, compute, cinder
  6. Deploy cluster
  7. Start ostf tests to check that changes do not reproduce regression

Change-Id: I4de4513e0f9186fda8750a74b44b5cff0edfbc21