Fix time sync after taking a snapshot
Currently time sync is only performed for the admin node, as an empty
list of nodes (default argument value) is passed to sync_time()
method.

On master branch this works as expected, as sync_time() behaviour
was changed in commit 03159d5cfb740ec1fb41420d3867a2477f31b2d1. The
latter seems to be a rather big change to be backported to the stable
branch, though. The alternative is to fix this in 8.0 only by passing
the list of nodes to sync_time().

No commit is done to master first as API of sync_time() changed
completely since 8.0 and those two would be entirely different
changes (and master defaults to all nodes of the cluster - so
we don't need to pass a list of nodes explicitly at all).

Closes-Bug: #1525214

Change-Id: Ibf639a5605f0ce29c9c3d3105b0c9c5880c8a965