Add smoke tests for new data-driven upgrade procedure
The upgrade procedure was completely changed because of changing base OS for Fuel (Centos7).
It leads to complete refactoring of all upgrade-related code including tests

This commit should be backportable to 7.0 branch because we need to prepare some 7.0 clusters.
This is the reason of rejecting SSHManager _for this commit_. It will be integrated later.

This commit includes base method for upgrading Fuel and introduces smoke tests for upgrading.

Change-Id: I1932c28beb9a7101019a02d4cd5245a1a0996a05
Partial-bug:1546943