Re-create ssh connection on revert_snapshot instead of re-using
This patch introduces workaround which re-creates
ssh connection instead of re-using one.
We need it in order not to stuck for 15 minutes each time after
resuming snapshot. So we decrease system test time on n*15,
where n - number of revert_snapshot operations and in the
most cases this value is 3. So we save 3*15 = 45 min for each
test

Related-bug: #1549274

Change-Id: I3b2f1a2058bbc736578ac3c30a92e938009ed597