Add wait into  _connect() check
We try to execute command 'cd~' on remote to check if connection is alive,
before start to use it. So sometimes if connection not isAlive we hangs
here for 15-16 minutes as well. As result  - we downgrate test execution time.

After moving this check in wait(), we check if we exit with succes
after 20 sec wait, and if not - TimeoutError will be raised.
It catched in follow except block and force reconnet() method.

Change-Id: If17b52071ba0d8eb163003e30fa835f57f1e4403
Related-Bug: 1549274