Fix inst.repo / inst.ks kernel parameters for Anaconda for USB boot
Adjusted as hd:LABEL:/ source are not suitable for installation from usb drive
because Anaconda search for ISO file with repo. Installation failed if it
can't find. So we change `repo` to cdrom:OpenStack_Fuel:/ and `ks` to
cdrom:OpenStack_Fuel:/ks.cfg both for cdrom and usb installation.

Also `repo` and `ks` kernel parameters changed to `inst.repo` and `inst.ks`

Change-Id: Icaf51417daa2426aa3714daa7a38d3c18b989de6
Related-Bug: #1540862