Working around some corner cases
The patch adds more checks to tests involved in reverting live
snapshots. This is needed because after reverting such a snapshot there
is a posibillity to have a broken RabbitMQ-cluster or Ceph installation.
The root cause of these problems is a time shift introduced by `ntpdate`
calls in revert_snapshot(). Increased the timeout in
assert_ha_services_ready().
Also added a function to restart Ceph services on all OSD-nodes
unconditionally. It only works for CentOS, as Ubuntu's Ceph has an
ability to restart its OSDs.

Change-Id: Ibc86080e46e079408b8302ad6d8b4d2664cdd38a
Closes-bug: #1558557
Closes-bug: #1557535
Signed-off-by: Denis V. Meltsaykin <dmeltsaykin@mirantis.com>