Working around some corner cases
The patch adds more checks to tests involved in reverting live
snapshots. This is needed because after reverting such a snapshot there
is a posibillity to have a broken RabbitMQ-cluster or Ceph installation.
The root cause of these problems is a time shift introduced by `ntpdate`
calls in revert_snapshot(). Increased the timeout in
assert_ha_services_ready().
Also added the repeat_until() function to make sure that HA services
are really UP, as there is a chance that pacemaker will restart the
RabbitMQ cluster after successful check due to bigger timeouts and new
logic.

Closes-bug: #1558557
Closes-bug: #1557535

Signed-off-by: Denis V. Meltsaykin <dmeltsaykin@mirantis.com>
Change-Id: Ibc86080e46e079408b8302ad6d8b4d2664cdd38a