Increase timeout to avoid test failure
In the method change_pacemaker_parameter_not_break_rabbitmq,
the timeout for the count_run_rabbit is too short.
That causes sometimes tests to fail with the AssertionError:
No Running rabbitmq nodes found.
In this commit, I increase the timeout and the interval to
insure that we don't get that test failure.

Change-Id: I73fe407078121f8ef4d46120299c86e45c879009
Closes-Bug: #1566324