Started refactoring helpers to use SSHManager
This is a first commit in series, that should
eventually refactor whole fuel-qa to use SSHManager class
for SSH access.

This commit addresses the usage of "get remote"-style
expressions in helper modules. However, due to a large scope
even in helpers alone, it doesn't fixes all the places
where SSHManager should be used.

Change-Id: I527efaef0874d8eb80b7f93b32ea52224a746593
Partial-Bug: #1566377
Partial-Bug: #1580324