Specify one more group for detach-rabbitmq tests for the plugin CI
We want to test detach-rabbitmq on Plugin CI. To do that, we need to
outline a set of tests we want to run on each commit. That is done with
the new 'detach-rabbitmq-ci-group' group, covering a couple of already
existing test cases. The group is intended for Plugin CI use only.

Change-Id: Ia75a39f1a3e889f4635909fdb52d993fcb11974d