Fix UCA package check function
This change fixes the dpkg-query parameters that get passed. The
\n was being interpreted as an actual line break rather than
being passed to the command line as a \n.  dpkg-query accepts the string
'\n' in the format to ensure the output has a line return in it. Because
the previous version was actually being interpreted as a line break, it
was causing errors in the ssh command when it was executed on the remote
system.

Change-Id: I1955e8809f851f8602991def6bb3fdad63cd6693
Closes-Bug: #1586170