Fix UCA tests
This change includes several string escaping fixes for the commands
used to check the source of the OpenStack packages.

This change fixes the dpkg-query parameters that get passed. The
'\n' was being interpreted as an actual line break rather than
being passed to the command line as '\n'.  dpkg-query accepts the string
'\n' in the format to ensure the output has a line return in it. Because
the previous version was actually being interpreted as a line break, it
was causing errors in the ssh command when it was executed on the remote
system.  Additionally this adjust the origin check to only check for the
UCA uri as the package may come from proposed or main.

Also included in this patch is the exclusion of using the glance
packages for checks as the murano-glance-artifacts-plugin package
is currently installed by Fuel but is a MOS specific package.

Change-Id: I1955e8809f851f8602991def6bb3fdad63cd6693
Closes-Bug: #1586170