[TestRail] Improve regexp for bug links
When the link to bug is added to result first time
(manually) it can have the following format:

https://launchpad.net/bugs/1549409

Currently used regex in statistics generator script
doesn't detect IDs in such links, so it skips them.

Hovewer, when the link is copied from previous result
by reporter, it's always converted to this format:

https://bugs.launchpad.net/fuel/+bug/1549409

So the issue with regex affects only very rare cases.

Change-Id: I3a7357b969682009d1e68fd821c0778e8a9eacb8