Fix instance volume mount script creation
Now the volume mount script would be created with correct newlines.
Also, this change contain some additional debugging information, so
the process of creating and mounting the volume logged to the
instance /var/log/messages through syslog. Also, this commit contain
the the workaround for the AuthenticationException because the
Cirros image starts ssh servece just before the cloud-init and in
some cases we need to wait when the instance_initial_scenario canges
the password for the cirros user.

Change-Id: Icb981c90a5ed8beb75fa119158161bf5ea3e5900
Closes-Bug: #1588808
Signed-off-by: Maksim Malchuk <mmalchuk@mirantis.com>