Pass retry count=1 into keystone session
KeystoneAuth can handle connection failure and timeouts not related to
target service but but defauit this retries are disabled. We are faced with
several issues related to broken session in case of changed time on server or
even restart of nailgun. The second request will re-create session inside of
KeystoneAuth so lets allow session to handle it.

Related-bug:1606887
Change-Id: I3011e25a5c81f74015430b5ace5aa14bc4695bb8