Allow getting release id by openstack version field
In upgrade chain scenario we can stuck in release fetching
because after restore of master node we have 3 releases which contains
"Ubuntu 14.04" substrings.
In previous implementation the first matched release is taken (which is Kilo
in upgrade 7-8-9 case instead of Mitaka)
Reusing of RELEASE_VERSION allow us to tune selection process.

Change-Id: Iac2b15fba774b3055016a67528528d8b851dfe16
Partial-bug:1620978