Do not expect RabbitMQ to be down on Pacemaker parameter change
Right now the test expects that all RabbitMQ nodes will be down
after parameter change. That is not always the case and moreover we
don't really care if it restarts or not. What we need instead is to
verify that cluster either does not break at all, or is fixed after
some time. So now we wait for a minute to give Pacemaker time to
shut down some nodes if it decides to do so and after that we wait
for RabbitMQ cluster to come up with timeout.

Closes-Bug: #1618909
Change-Id: I50f65f73c6937d23721c9e236f2a8059313e760f