Add extra factor of TC definition in testrail
We need to execute the upgrade tests for different chains of
MOS versions. And for the each run we need to store and process
test result in Testrail, but there is problem - the upgrade tests
for various MOS version have the same test group and for now
testrail reporter cannot tell apart them correctly, because
reporter does filter tests by test group parameter. So, we need
to add additional factor of Test Case definition into test report
mechanism

Change-Id: I1f4458a387f7dd9e8061a42023d1cb94c7028512