Fix plugin v3 and v4 hotplug tests
The example plugin v3 has not legacy tasks
that's why propagate_task_deploy should not
be introduced.

The plugin v4 should be installed at the
new introduced by plugin role fuel_plugin_example_v4
and should be checked under that role.
The controller role is not affected by plugin v4,
so, it was decided to disable 2 tests in which
we re-deploy controller node.

Change-Id: I13eb7d7189421b8ae82410866440cba2052fac69
Closes-Bug:1620325