Increase reboot_plugin timeout
It looks like recent changes increased the time needed to reboot
the node using plugin's reboot task. I was able to reproduce it
in my lab, so it is a common issue.

An obvious way to fix this problem is to increase reboot timeout.
I have set it to 10 mins to overcome all possible issues.

Change-Id: I327087a050a475c607ed89e226d0f2e5355e2c45
Closes-bug: #1617329