[8.0] Add a reboot procedure for updates
Since sometimes we update kernels and system libraries for a master node
we have to reboot it after an update in order changes to take effect. This
commit adds simplified mechanics for rebooting of the master node.
Also refactored the order of actions in packages updates, as it is
better to run an update having containers destroyed; removed the call to
"sync deployment tasks" as it is already contained in the fuel-library
package.
Added PEP8 conformance.

Change-Id: I9ad9201e0010763f1668ef04e8732ece69e9dc71