Reset aux master node repo priority according to env config
We need to be able to set aux master repo priority to avoid
issues when we have a repo preconfigured already, but we add
some additional repos with other priorities and cannot
control things and get packages excluded.
This commit adds 2 additional settings that specify aux repo
name and its priority.

Change-Id: I6da435c393cbead6fbfca0421c05de399471137a
Related-bug: #1658063