Add timeout to check_service_was_restated
There is the function to check that service was restarted. And there is
test case two_clusters_reconfiguration, which uploads openstack-config
with
  notify: Service[nova-scheduler]
and checks that puppet restarts the nova-scheduler.

Unfortunately, there is a time gap between uploading config to nailgun
and applying puppet on a node. So we have to add a timeout to the
checker for covering that case.

Closes-bug: #1673750
Change-Id: Icef1e15489727a13a753b9921df0cc5527f9ab2d