[8.0] Increase timeout waiting for MySQL to be ready
In a scenario with hard reset of a controller node there is a
probabillity that MySQL Galera will take much more time to
repair itself than the usual 4 minutes. Thus, this commit
increases the timeout up to 10 minutes.

While here adjust the timeout of node readiness up to 5 minutes.

Change-Id: I830681150281d9da5cddd66f5c5b758076308b71
Related-Bug: #1507731